subscriptionEventsPosted and subscriptionEventsProcessed need to be unsigned long.
Manifests in gateway which assumes these counters roll over on 64bits
the same as all other gateway counters.
Without this fix the gateway serverEventRate and serverPostRate PVs
spike to 1.8e18 each time the underlying counters roll over 32bits.

+++ src/pcas/generic/caServer.cc
@@ -154,17 +154,17 @@ class epicsTimer & caServer::createTimer ()
     return fileDescriptorManager.createTimer ();
 }
 
-unsigned caServer::subscriptionEventsProcessed () const
+unsigned long caServer::subscriptionEventsProcessed () const
 {
     if ( pCAS ) {
         return this->pCAS->subscriptionEventsProcessed ();
     }
     else {
-        return 0u;
+        return 0UL;
     }
 }
 
-unsigned caServer::subscriptionEventsPosted () const
+unsigned long caServer::subscriptionEventsPosted () const
 {
     if ( pCAS ) {
         return this->pCAS->subscriptionEventsPosted ();
